Indonesian Translation (#114)
Some translation is kind of similar to another line (like on line 411 with 403, because Indonesia is kinda having a lack in word richness(?)). Also, I'm quite not sure about some line because I'm using English on any device as the main language. Maybe I will fix the translation if I made a mistake(s).